Many updates including plain text input. Not tested with different display text yet

+++ carbide_core/src/text/types/font_weight.rs
@@ -31,18 +31,18 @@ pub enum FontWeight {
 }
 
 impl FontWeight {
-    pub fn weight(&self) -> f64 {
+    pub fn weight(&self) -> u16 {
         match self {
-            FontWeight::W100 | FontWeight::Thin => 100.0,
-            FontWeight::W200 | FontWeight::ExtraLight => 200.0,
-            FontWeight::W300 | FontWeight::Light => 300.0,
-            FontWeight::W400 | FontWeight::Normal => 400.0,
-            FontWeight::W500 | FontWeight::Medium => 500.0,
-            FontWeight::W600 | FontWeight::SemiBold => 600.0,
-            FontWeight::W700 | FontWeight::Bold => 700.0,
-            FontWeight::W800 | FontWeight::ExtraBold => 800.0,
-            FontWeight::W900 | FontWeight::Black => 900.0,
-            FontWeight::Other(val) => *val as f64,
+            FontWeight::W100 | FontWeight::Thin => 100,
+            FontWeight::W200 | FontWeight::ExtraLight => 200,
+            FontWeight::W300 | FontWeight::Light => 300,
+            FontWeight::W400 | FontWeight::Normal => 400,
+            FontWeight::W500 | FontWeight::Medium => 500,
+            FontWeight::W600 | FontWeight::SemiBold => 600,
+            FontWeight::W700 | FontWeight::Bold => 700,
+            FontWeight::W800 | FontWeight::ExtraBold => 800,
+            FontWeight::W900 | FontWeight::Black => 900,
+            FontWeight::Other(val) => *val,
         }
     }
 }
